Improve mock helper signatures
