+++ Tests/PlayerTests/NowPlayingInfoMetadataPublisherTests.swift
@@ -30,7 +30,7 @@ final class NowPlayingInfoMetadataPublisherTests: TestCase {
 
     func testAvailableAfterDelay() {
         let player = Player(
-            item: .asynchronous(url: Stream.onDemand.url, metadata: LocalMetadata(title: "title"), after: 0.5)
+            item: .mock(url: Stream.onDemand.url, loadedAfter: 0.5, withMetadata: AssetMetadataMock(title: "title"))
         )
         expectAtLeastSimilarPublished(
             values: [[:], [MPMediaItemPropertyTitle: "title"]],
@@ -41,7 +41,7 @@ final class NowPlayingInfoMetadataPublisherTests: TestCase {
     func testImmediatelyAvailableWithMetadata() {
         let player = Player(item: .simple(
             url: Stream.onDemand.url,
-            metadata: LocalMetadata(
+            metadata: AssetMetadataMock(
                 title: "title",
                 subtitle: "subtitle",
                 description: "description"
@@ -60,7 +60,7 @@ final class NowPlayingInfoMetadataPublisherTests: TestCase {
     }
 
     func testUpdate() {
-        let player = Player(item: .updated(url: Stream.onDemand.url, after: 1))
+        let player = Player(item: .mock(url: Stream.onDemand.url, withMetadataUpdateAfter: 1))
         expectAtLeastSimilarPublished(
             values: [
                 [
@@ -79,7 +79,7 @@ final class NowPlayingInfoMetadataPublisherTests: TestCase {
     }
 
     func testNetworkItemReloading() {
-        let player = Player(item: .networkLoaded(url: Stream.onDemand.url, media: .media1))
+        let player = Player(item: .mock(media: .media1))
         expectAtLeastSimilarPublished(
             values: [
                 [:],
@@ -103,12 +103,12 @@ final class NowPlayingInfoMetadataPublisherTests: TestCase {
             from: player.nowPlayingInfoMetadataPublisher()
         ) {
             player.removeAllItems()
-            player.append(.networkLoaded(url: Stream.onDemand.url, media: .media1))
+            player.append(.mock(media: .media1))
         }
     }
 
     func testEntirePlayback() {
-        let player = Player(item: .simple(url: Stream.shortOnDemand.url, metadata: LocalMetadata(title: "title")))
+        let player = Player(item: .simple(url: Stream.shortOnDemand.url, metadata: AssetMetadataMock(title: "title")))
         expectAtLeastSimilarPublished(
             values: [[MPMediaItemPropertyTitle: "title"], [:]],
             from: player.nowPlayingInfoMetadataPublisher()
